Refactor & UI Changes
* Refactor code to split it up a bit more into specific classes & rename consistently
* Add extra supported parameters to generate in Pipeline
* Update UI in TextoImageView to handle all values to be supported (but some are not usable currently)
